requirements.txt
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.6.3
timm==0.4.12
sklearn 
tqdm
pillow
matplotlib
numpy
intern
umap-learn
segmentation-models-pytorch==0.3.0-dev
torchsummary
torchsummaryx
rich
opencv-python
einops
kornia==0.6.0
connected-components-3d
tensorboard
networkx
pyrsistent
scikit-image
pandas
